moved interface formulation to Interface
