+++ src/festim/__init__.py
@@ -52,7 +52,6 @@ from .hydrogen_transport_problem import (
     HydrogenTransportProblem,
     HydrogenTransportProblemDiscontinuous,
     HydrogenTransportProblemDiscontinuousChangeVar,
-    InterfaceMethod,
 )
 from .initial_condition import (
     InitialConcentration,
@@ -70,7 +69,7 @@ from .settings import Settings
 from .source import HeatSource, ParticleSource, SourceBase
 from .species import ImplicitSpecies, Species, find_species_from_name
 from .stepsize import Stepsize
-from .subdomain.interface import Interface
+from .subdomain.interface import Interface, InterfaceMethod
 from .subdomain.surface_subdomain import (
     SurfaceSubdomain,
     SurfaceSubdomain1D,
