+++ src/festim/subdomain/interface.py
@@ -1,9 +1,35 @@
+from enum import Enum
+from typing import TYPE_CHECKING
+
 import dolfinx
 import numpy as np
+import ufl
 from dolfinx.cpp.fem import compute_integration_domains
 from packaging.version import Version
 
+from festim.material import SolubilityLaw
 from festim.subdomain.volume_subdomain import VolumeSubdomain
+
+if TYPE_CHECKING:
+    from festim.species import Species
+
+
+class InterfaceMethod(Enum):
+    """How to couple interfaces for discontinuous problems."""
+
+    nitsche = 10
+    penalty = 20
+
+    @classmethod
+    def from_string(cls, s: str):
+        """Can be removed with Python 3.11+."""
+        s = s.lower()
+        if s == "nitsche":
+            return cls.nitsche
+        elif s == "penalty":
+            return cls.penalty
+        else:
+            raise ValueError("interface_method must be one of 'nitsche' or 'penalty'")
 
 
 class Interface:
@@ -58,7 +84,8 @@ class Interface:
 
         integration_data = compute_integration_domains(*args).reshape(-1, 4)
         for i in range(2):
-            # We pad the parent to submesh map to make sure that sparsity pattern is correct
+            # We pad the parent-to-submesh map so that the sparsity pattern
+            # is correct.
             mapped_cell_0 = self.subdomains[i].parent_to_submesh[integration_data[:, 0]]
             mapped_cell_1 = self.subdomains[i].parent_to_submesh[integration_data[:, 2]]
             max_cells = np.maximum(mapped_cell_0, mapped_cell_1)
@@ -133,3 +160,126 @@ class Interface:
         assert (domain1_cell >= 0).all()
 
         return (self.id, ordered_integration_data.reshape(-1))
+
+    # TODO this should be a method of a subclass of Interface since we want to support
+    # other types of interfaces in the future
+    def set_formulation(
+        self, dInterface, method: InterfaceMethod, species: list["Species"], temperature
+    ):
+
+        subdomain_0, subdomain_1 = self.subdomains
+        res = self.restriction
+        mesh = dInterface.ufl_domain()
+
+        for spe in species:
+            assert subdomain_0 in spe.subdomains and subdomain_1 in spe.subdomains, (
+                f"Species {spe.name} must be defined in both subdomains of the "
+                "interface for the interface conditions to be applied"
+            )
+            v_b = spe.subdomain_to_test_function[subdomain_0](res[0])
+            v_t = spe.subdomain_to_test_function[subdomain_1](res[1])
+
+            u_b = spe.subdomain_to_solution[subdomain_0](res[0])
+            u_t = spe.subdomain_to_solution[subdomain_1](res[1])
+
+            K_b = subdomain_0.material.get_solubility_coefficient(
+                mesh, temperature(res[0]), spe
+            )
+            K_t = subdomain_1.material.get_solubility_coefficient(
+                mesh, temperature(res[1]), spe
+            )
+
+            match method:
+                case InterfaceMethod.penalty:
+                    F_0, F_1 = self.penalty_method(
+                        dInterface, (u_b, u_t), (K_b, K_t), (v_b, v_t)
+                    )
+
+                case InterfaceMethod.nitsche:
+                    F_0, F_1 = self.nitsche_method(
+                        dInterface, (u_b, u_t), (K_b, K_t), (v_b, v_t)
+                    )
+
+                case _:
+                    raise ValueError(
+                        f"Unknown interface method {self.method_interface}"
+                    )
+            subdomain_0.F += F_0
+            subdomain_1.F += F_1
+
+    def penalty_method(self, dInterface, us, Ks, vs):
+        subdomain_0, subdomain_1 = self.subdomains
+        u_0, u_1 = us
+        v_0, v_1 = vs
+        K_0, K_1 = Ks
+        if subdomain_0.material.solubility_law == subdomain_1.material.solubility_law:
+            left = u_0 / K_0
+            right = u_1 / K_1
+        else:
+            match subdomain_0.material.solubility_law:
+                case SolubilityLaw.HENRY:
+                    left = u_0 / K_0
+                case SolubilityLaw.SIEVERT:
+                    left = (u_0 / K_0) ** 2
+                case _:
+                    raise ValueError(
+                        "Unsupported material law "
+                        + f"{subdomain_0.material.solubility_law}"
+                    )
+
+            match subdomain_1.material.solubility_law:
+                case SolubilityLaw.HENRY:
+                    right = u_1 / K_1
+                case SolubilityLaw.SIEVERT:
+                    right = (u_1 / K_1) ** 2
+                case _:
+                    raise ValueError(
+                        f"Unsupported material law "
+                        f"{subdomain_1.material.solubility_law}"
+                    )
+
+        equality = right - left
+
+        F_0 = self.penalty_term * ufl.inner(equality, v_0) * dInterface(self.id)
+        F_1 = -self.penalty_term * ufl.inner(equality, v_1) * dInterface(self.id)
+
+        return F_0, F_1
+
+    def nitsche_method(self, dInterface, us, Ks, vs):
+        u_b, u_t = us
+        K_b, K_t = Ks
+        v_b, v_t = vs
+
+        def mixed_term(u, v, n):
+            return ufl.dot(ufl.grad(u), n) * v
+
+        res = self.restriction
+        n = ufl.FacetNormal(dInterface.ufl_domain())
+        cr = ufl.Circumradius(dInterface.ufl_domain())
+        n_0 = n(res[0])
+        h_0 = 2 * cr(res[0])
+        h_1 = 2 * cr(res[1])
+        gamma = self.penalty_term
+        F_0 = -0.5 * mixed_term((u_b + u_t), v_b, n_0) * dInterface(
+            self.id
+        ) - 0.5 * mixed_term(v_b, (u_b / K_b - u_t / K_t), n_0) * dInterface(self.id)
+
+        F_1 = +0.5 * mixed_term((u_b + u_t), v_t, n_0) * dInterface(
+            self.id
+        ) - 0.5 * mixed_term(v_t, (u_b / K_b - u_t / K_t), n_0) * dInterface(self.id)
+        F_0 += (
+            2
+            * gamma
+            / (h_0 + h_1)
+            * (u_b / K_b - u_t / K_t)
+            * v_b
+            * dInterface(self.id)
+        )
+        F_1 += (
+            -2
+            * gamma
+            / (h_0 + h_1)
+            * (u_b / K_b - u_t / K_t)
+            * v_t
+            * dInterface(self.id)
+        )
